Refine team directory modules and CI baseline
